Correct # of permutations
You can't get 5! because changing which state is the start state changes the behavior of the machine. But you can permute the remaining 4! states (and 2 dirs).


Actually, technically, it's even possible to specify machines where not all 4! permutations are unique, like: `1RB ---  1RC 0LB  1RB 0LC` which is identical when permuting states B & C ... but perhaps this is getting too pedantic :/